requirements.txt
Django>=5.2.1
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
Pillow>=10.0.0  # для работы с изображениями
python-telegram-bot>=20.0 